power: pm8921-charger: fix race in unplug check worker
When the usb voltage collapses, the driver waits for 50ms and checks
whether the usb voltage has restored. If it has, the driver treats
it as collapse and resets the usb current and the target
current. It assumes that unplug check worker will run and do its
aicl duties to step up the usb current.

However if the unplug check worker runs right at the time when usb
has collapsed it will exit and wont schedule anymore.

To fix this check if the unplug worker is pending, if not, reschedule
it.

(cherry picked from commit 6bf4d14dc4ff9b6ac0b558af599a23d82909badf)

Change-Id: I74cce5edd10fd6fb7b357df2a31de5b3dbaeb4b9
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>