media: dvb: Fix race condition in DVR thread
When DVR thread is created, it may happen that user sets
DVR buffer mode to external buffer but did not yet set the
buffer itself, meanwhile, DVR thread checks whether the
input buffer was already set and if not the thread terminates.

This causes termination of DVR thread for no good reason.
DVR thread should be terminated only when DVR device is closed.
The change removes termination of the thread when input buffer
is not set yet to avoid such race condition.

CRs-Fixed: 493761
Change-Id: I57c9acd13bd493ae08d28703bf5f6c0dea55c20a
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>